core, configs: ivshmem: Add support for up to 16 MSI-X vectors
This allows ivshmem devices with MSI-X support to expose up to 16
(PCI_EMBEDDED_MSIX_VECTS) vectors to its users. The doorbell register is
now interpreting the written value as vector number. It is possible to
restrict the number of available vectors via the cell configuration. Out
of range writes to the doorbell register will simply be ignored.

In case the cell uses less than the exposed vectors, it is now important
to correctly mask them initially in the MSI-X table as the PCI spec
demands, or the hypervisor may find invalid vectors and complain.

With the number of vectors increasing, we also need to adjust the size
of the related resource region in the cell configs accordingly.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>